+++ lib/lite/query.rb
@@ -4,5 +4,4 @@
   require "lite/query/#{name}"
 end
 
-require 'generators/lite/query/install_generator'
 require 'generators/rails/query_generator'