+++ spec/generators/rails/query_generator_spec.rb
@@ -11,7 +11,6 @@ RSpec.describe Rails::QueryGenerator, type: :generator do
   end
 
   let(:query_path) { 'spec/generators/tmp/app/queries/v1/users/age_query.rb' }
-  let(:rspec_path) { 'spec/generators/tmp/spec/queries/v1/users/age_query_spec.rb' }
 
   describe '#generator' do
     context 'when generating app files' do
@@ -21,22 +20,9 @@ RSpec.describe Rails::QueryGenerator, type: :generator do
 
       it 'to include the proper markup' do
         query_file = File.read(query_path)
-        text_snippet = 'class V1::Users::AgeQuery < ApplicationQuery'
+        text_snippet = 'class V1::Users::AgeQuery < Lite::Query::Base'
 
         expect(query_file.include?(text_snippet)).to eq(true)
-      end
-    end
-
-    context 'when generating spec files' do
-      it 'to be true when query file exists' do
-        expect(File.exist?(rspec_path)).to eq(true)
-      end
-
-      it 'to include the proper markup' do
-        rspec_file = File.read(rspec_path)
-        text_snippet = 'V1::Users::AgeQuery, type: :query'
-
-        expect(rspec_file.include?(text_snippet)).to eq(true)
       end
     end
   end